The path finder project is now more successful. It currently finds paths under a wider variety of environments, and I've made some improvements to prevent diagonal movement unless it is necessary.

+++ trunk/AS2/Structs.cs
@@ -13,9 +13,9 @@ namespace AS2
             public int X;
             public int Y;
             public int speed;
-            public int heuristic;
-            public int G;
-            public int F;
+            public float heuristic;
+            public float G;
+            public float F;
             public int parentID;
             public bool walkable;
         }
